LayoutTests/storage/indexeddb/intversion-close-in-upgradeneeded-expected.txt

Test that when db.close is called in upgradeneeded, the db is cleaned up on refresh.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-close-in-upgradeneeded.html"
indexedDB.deleteDatabase(dbname)
request = indexedDB.open(dbname, 7)


openSuccess():
FAIL sawTransactionComplete should be true. Was false.
db = event.target.result
FAIL db.version should be 7 (of type number). Was  (of type string).
transaction = db.transaction('os')
FAIL transaction = db.transaction('os') threw exception Error: NotFoundError: DOM IDBDatabase Exception 8
PASS successfullyParsed is true

TEST COMPLETE
